made Router component
