+++ src/index.tsx
@@ -3,14 +3,14 @@ import ReactDOM from 'react-dom/client'
 import { Provider } from 'react-redux'
 import { BrowserRouter } from 'react-router-dom'
 
+// Common styles
+import './scss/index.scss'
+
 // Store
 import { store } from 'redux/store'
 
 // Components
 import { App } from 'components'
-
-// Styles
-import './scss/index.scss'
 
 // Tests
 import reportWebVitals from './reportWebVitals'
